moved Processor and its interface to their own namespace, also moved SerializerInterface to its namespace
