+++ src/ProcessorInterface.php
@@ -1,17 +1,16 @@
 <?php
 namespace Thunder\Shortcode;
 
+use Thunder\Shortcode\Processor\ProcessorInterface as ProcessorInterfaceBase;
+
 /**
- * @author Tomasz Kowalczyk <tomasz@kowalczyk.cc>
+ * This implementation is left only to not break IDE autocompletion, this class
+ * is deprecated, it was moved to the new location as specified in docblock.
+ * This file will be removed in version 1.0!
+ *
+ * @deprecated use Thunder\Shortcode\Extractor\RegexExtractor
+ * @codeCoverageIgnore
  */
-interface ProcessorInterface
+interface ProcessorInterface extends ProcessorInterfaceBase
     {
-    /**
-     * Process text using registered shortcode handlers
-     *
-     * @param string $text
-     *
-     * @return string
-     */
-    public function process($text);
     }
